file structure changed, currentWeatherPojo and injector moved to correct file location
