--- src/test/java/org/pojo/CurrentWeatherPojo.java
@@ -1,134 +1,0 @@
-package org.pojo;
-
-import java.util.List;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class CurrentWeatherPojo {
-
-	@JsonProperty("rain")
-	private Rain rain;
-
-	@JsonProperty("visibility")
-	private Integer visibility;
-
-	@JsonProperty("timezone")
-	private Integer timezone;
-
-	@JsonProperty("main")
-	private Main main;
-
-	@JsonProperty("clouds")
-	private Clouds clouds;
-
-	@JsonProperty("sys")
-	private Sys sys;
-
-	@JsonProperty("dt")
-	private Integer dt;
-
-	@JsonProperty("coord")
-	private Coord coord;
-
-	@JsonProperty("snow")
-	private Snow snow;
-
-	@JsonProperty("weather")
-	private List<WeatherItem> weather;
-
-	@JsonProperty("name")
-	private String name;
-
-	@JsonProperty("cod")
-	private Integer cod;
-
-	@JsonProperty("id")
-	private Integer id;
-
-	@JsonProperty("base")
-	private String base;
-
-	@JsonProperty("wind")
-	private Wind wind;
-
-	public Rain getRain(){
-		return rain;
-	}
-
-	public Integer getVisibility(){
-		return visibility;
-	}
-
-	public Integer getTimezone(){
-		return timezone;
-	}
-
-	public Main getMain(){
-		return main;
-	}
-
-	public Clouds getClouds(){
-		return clouds;
-	}
-
-	public Sys getSys(){
-		return sys;
-	}
-
-	public Integer getDt(){
-		return dt;
-	}
-
-	public Coord getCoord(){
-		return coord;
-	}
-
-	public Snow getSnow(){
-		return snow;
-	}
-
-	public List<WeatherItem> getWeather(){
-		return weather;
-	}
-
-	public String getName(){
-		return name;
-	}
-
-	public Integer getCod(){
-		return cod;
-	}
-
-	public Integer getId(){
-		return id;
-	}
-
-	public String getBase(){
-		return base;
-	}
-
-	public Wind getWind(){
-		return wind;
-	}
-
-	@Override
- 	public String toString(){
-		return 
-			"Response{" + 
-			"rain = '" + rain + '\'' + 
-			",visibility = '" + visibility + '\'' + 
-			",timezone = '" + timezone + '\'' + 
-			",main = '" + main + '\'' + 
-			",clouds = '" + clouds + '\'' + 
-			",sys = '" + sys + '\'' + 
-			",dt = '" + dt + '\'' + 
-			",coord = '" + coord + '\'' + 
-			",snow = '" + snow + '\'' + 
-			",weather = '" + weather + '\'' + 
-			",name = '" + name + '\'' + 
-			",cod = '" + cod + '\'' + 
-			",id = '" + id + '\'' + 
-			",base = '" + base + '\'' + 
-			",wind = '" + wind + '\'' + 
-			"}";
-		}
-}
